+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## [v0.0.1] - 2023-06-20
+## [v0.1.0] - 2023-07-17
+
+### Added
+
+- Pinch and rotate support.
+- Better README.
+- Covered with tests.
+
+## [v0.0.1.alpha] - 2023-06-20
 
 ### Added